requirements.txt
pandas
langchain
langchain-openai
langchain_experimental
tabulate
ipython
python-dotenv 